MCP_chat_wether/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "MCP Weather Israel with Playwright and Google Gemini"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.3.0",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
    "playwright>=1.44.0",
]
